REL: rename the library
