+++ dask_streams/tests/test_core.py
@@ -9,11 +9,11 @@ from tornado import gen
 from tornado.queues import Queue
 from tornado.ioloop import IOLoop
 
-import streams as s
+import dask_streams as s
 
 from ..core import Stream
-from streams.sources import sink_to_file, Counter
-from streams.utils_test import inc, double, gen_test, tmpfile
+from dask_streams.sources import sink_to_file, Counter
+from dask_streams.utils_test import inc, double, gen_test, tmpfile
 
 
 def test_basic():
