+++ examples/plot_graph.py
@@ -1,4 +1,4 @@
-from streams import Stream
+from dask_streams import Stream
 from operator import add
 
 source1 = Stream(name='source1')
